// src/shared/api/generated/models/supportInternalNotePageResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { SupportInternalNoteResponseDto } from "./supportInternalNoteResponseDto";

export interface SupportInternalNotePageResponseDto {
  items: SupportInternalNoteResponseDto[];
  /** @nullable */
  nextCursor: string | null;
}